requirements.txt
google-generativeai==0.5.0
pandas==2.2.2
gspread==6.0.2
reportlab==4.1.0
python-dotenv==1.0.1
whisper-timestamped==1.15.8
qrcode[pil]==7.4.2
Flask==3.0.3
google-cloud-speech==2.24.0  # For audio transcription
html2text
apscheduler
tenacity